+++ core/src/main/scala/ai/lum/odinson/lucene/analysis/NormalizedTokenStream.scala
@@ -1,12 +1,13 @@
 package ai.lum.odinson.lucene.analysis
 
-import com.ibm.icu.text.Normalizer2
 import org.apache.lucene.analysis.TokenStream
 import org.apache.lucene.analysis.tokenattributes.{ CharTermAttribute, PositionIncrementAttribute }
+import ai.lum.odinson.utils.Normalizer
 
-class NormalizedTokenStream(val tokenSeqs: Seq[Seq[String]]) extends TokenStream {
-
-  val normalizer = Normalizer2.getNFKCCasefoldInstance()
+class NormalizedTokenStream(
+  val tokenSeqs: Seq[Seq[String]],
+  val aggressive: Boolean = false,
+) extends TokenStream {
 
   val posIncrAtt = addAttribute(classOf[PositionIncrementAttribute])
   val termAtt = addAttribute(classOf[CharTermAttribute])
@@ -23,7 +24,7 @@ class NormalizedTokenStream(val tokenSeqs: Seq[Seq[String]]) extends TokenStream
   private def mkSynonyms(tokenSeqs: Seq[Seq[String]]): Vector[Vector[String]] = {
     val synonyms = for (i <- tokenSeqs.head.indices) yield {
       tokenSeqs
-        .map(tokens => normalizer.normalize(tokens(i)))
+        .map(tokens => Normalizer.normalize(tokens(i), aggressive))
         .distinct
         .toVector
     }