+++ CHANGES.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- All strings are normalized with NFC, except the norm field which uses NFKC with casefolding and diacritic stripping.
+  This is the case both at index time and query time. This means you should reindex
+  if you upgrade to this version.
 ### Added
 - Added `ExtractorEngine.inMemory(...)` to help build an index in memory.